backend/services/xclip_service/requirements.txt
accelerate==1.12.0
torch==2.9.1
transformers==4.57.3
litserve==0.2.5
python-dotenv==1.2.1
python-logging-loki==0.3.1
python-multipart==0.0.20
pillow==12.0.0
bitsandbytes==0.49.1
